alarme feito

+++ Servidor/main.c
@@ -48,8 +48,7 @@ CLIENTES clientes;
 RESULTADOS resultados;
 int Ndefesa, Navanc, sair = 0;
 
-pthread_t jogo, tempo, tarefa_bola; 
-
+pthread_t jogo, tempo, tarefa_bola;
 
 /*
  * 0 - NAO OCUPADO
@@ -79,37 +78,35 @@ void bolaParaMeio() {
 
 void atualizaCampo(serv_clie * j) {
     int i, fd;
-    pthread_mutex_lock(&trinco);
+//        pthread_mutex_lock(&trinco);
 
     for (i = 0; i < clientes.tam; i++) {
-        if (clientes.c[i].logado) 
-        {
+        if (clientes.c[i].logado) {
             char str[80];
             sprintf(str, "/tmp/ccc%d", clientes.c[i].id);
-            
-            if (access(str, F_OK) != 0)
-            {
+
+            if (access(str, F_OK) != 0) {
                 resultados.numClientes--;
-                clientes.c[i].logado=0;
-                if(clientes.c[i].jogador!=NULL)
-                {
-                    clientes.c[i].jogador->humano=0;
-                    clientes.c[i].jogador=NULL;
+                clientes.c[i].logado = 0;
+                if (clientes.c[i].jogador != NULL) {
+                    clientes.c[i].jogador->humano = 0;
+                    clientes.c[i].jogador = NULL;
                 }
-                clientes.c[i].equi='-';
+                clientes.c[i].equi = '-';
                 continue;
             }
             fd = open(str, O_WRONLY);
             j->flag_campo = 1;
             j->flag_logado = 1;
             j->resultados = resultados;
+
             write(fd, j, sizeof (serv_clie));
 
             //printf("\n{Servidor} Dados enviados para o cliente %d.\n" "Tamanho pretendido: %d Tamanho enviado: %d \n Jogador enviado: %c", clientes.c[i].id, sizeof (serv_clie), a, j->jogador);
             close(fd);
         }
     }
-    pthread_mutex_unlock(&trinco);
+    //    pthread_mutex_unlock(&trinco);
 }
 
 void inicializaCampo() {
@@ -150,14 +147,14 @@ void inicializaCampo() {
     usleep(100);
 }
 
-void montaCampo(){
+void montaCampo() {
     int i = 0;
-    
+
     JOG[0][i].posicao.x = 11;
     JOG[0][i].posicao.y = 1;
     JOG[1][i].posicao.x = 11;
     JOG[1][i].posicao.y = 50;
-    
+
     i++;
     switch (Ndefesa) {
         case 1:
@@ -235,7 +232,7 @@ void montaCampo(){
             i++;
             break;
     }
-    
+
     switch (Navanc) {
         case 1:
             JOG[0][i].posicao.x = 11;
@@ -324,7 +321,7 @@ void inicializaJogadores() {
     JOG[0][0].precisao_remate = 25;
     JOG[0][0].tempo = 300000;
     JOG[0][0].equi = 'a';
-    
+
 
     JOG[1][0].fim = 0;
     JOG[1][0].humano = 0;
@@ -356,7 +353,7 @@ void inicializaJogadores() {
         cont++;
     }
 
-    
+
 
     //---------------------
     //INICIAOLIZACAO AVANC
@@ -378,7 +375,7 @@ void inicializaJogadores() {
 
         cont++;
     }
-    
+
     //--------------------- 
 
     bolaParaMeio();
@@ -418,10 +415,10 @@ void verificaGolo() {
         for (i = 0; i < clientes.tam; i++) {
             clientes.c[i].jogador->humano = 0;
             clientes.c[i].jogador = NULL;
-            
+
 
             sprintf(str, "/tmp/ccc%d", clientes.c[i].id);
-            inicializaCampo(str);
+            //            inicializaCampo(str);
         }
     }
 }
@@ -485,7 +482,7 @@ void * bola(void * dados) {
             } else {
                 j.jogador = 'o' + 2;
                 ball.y = posse_bola->posicao.y - 1;
-                
+
             }
 
             ball.x = posse_bola->posicao.x;
@@ -503,7 +500,7 @@ void * bola(void * dados) {
 
             j.resultados = resultados;
 
-            atualizaCampo(&j);
+            //            atualizaCampo(&j);
         }
         //
         //            for(a=0;a<2;a++)
@@ -570,9 +567,11 @@ void * move_jogador(void * dados) {
 
             if (jog->posicao.y + d.y >= 0 && jog->posicao.y + d.y < MaxY &&
                     jog->posicao.x + d.x >= 0 && jog->posicao.x + d.x < MaxX) {
-
+                
+                pthread_mutex_lock(&trinco);
                 if (isOcupado(jog->posicao.x + d.x, jog->posicao.x + d.x) == 1)
                     continue;
+//                pthread_mutex_unlock(&trinco);
 
                 j.jogador = '0' + jog->num;
                 j.equipa = jog->equi;
@@ -586,6 +585,7 @@ void * move_jogador(void * dados) {
                 j.xnovo = jog->posicao.x;
                 j.ynovo = jog->posicao.y;
                 j.resultados = resultados;
+                        pthread_mutex_unlock(&trinco);
 
                 atualizaCampo(&j);
 
@@ -690,8 +690,10 @@ void * move_jogador(void * dados) {
 
             }
 
+            usleep(jog->tempo);
+
         }
-        usleep(jog->tempo);
+
     }
     pthread_exit(0);
 }
@@ -1368,7 +1370,7 @@ void * Func_receber_cliente(void * dados) {
             //                    close(fd);
             //                    continue;
             //                } else {
-            int aux=0, t;
+            int aux = 0, t;
             j.flag_logado = 0;
             j.flag_campo = 0;
 
@@ -1413,25 +1415,21 @@ void * Func_receber_cliente(void * dados) {
 
             sleep(1);
             if (!resultados.fim)
-                inicializaCampo(str);
+                //                inicializaCampo(str);
 
 
-            fclose(f);
-        } 
-        else if (cliente.flag_desliga) 
-        {
-            for (i = 0; i < clientes.tam; i++) 
-            {
-                if (clientes.c[i].id == cliente.id) 
-                {
+                fclose(f);
+        }
+        else if (cliente.flag_desliga) {
+            for (i = 0; i < clientes.tam; i++) {
+                if (clientes.c[i].id == cliente.id) {
                     clientes.c[i].logado = 0;
                     strcpy(clientes.c[i].username, " ");
-                    clientes.c[i].equi='-';
-                    if(clientes.c[i].jogador!=NULL)
-                    {
-                        clientes.c[i].jogador->humano=0;
-                        clientes.c[i].jogador=NULL;
-                    } 
+                    clientes.c[i].equi = '-';
+                    if (clientes.c[i].jogador != NULL) {
+                        clientes.c[i].jogador->humano = 0;
+                        clientes.c[i].jogador = NULL;
+                    }
                     break;
                 }
             }
@@ -1464,26 +1462,26 @@ void * Func_receber_cliente(void * dados) {
 void comecaJogo() {
     int i;
 
+    montaCampo();
     resultados.res_eq1 = 0;
     resultados.res_eq2 = 0;
     resultados.fim = 0;
 
     pthread_create(&tarefa_bola, NULL, &bola, NULL);
 
-    pthread_create(&JOG[0][0].thread, NULL, &move_redes, (void *) &JOG[0][0]);
-    pthread_create(&JOG[0][6].thread, NULL, &move_jogador, (void *) &JOG[0][6]);
-    /*pthread_create(&JOG[1][0].thread, NULL, &move_redes, (void *) &JOG[1][0]);
+    //    pthread_create(&JOG[0][0].thread, NULL, &move_redes, (void *) &JOG[0][0]);
+    //    pthread_create(&JOG[1][0].thread, NULL, &move_redes, (void *) &JOG[1][0]);
 
     for (i = 1; i < TOTAL; i++) {
         pthread_create(&JOG[0][i].thread, NULL, &move_jogador, (void *) &JOG[0][i]);
         pthread_create(&JOG[1][i].thread, NULL, &move_jogador, (void *) &JOG[1][i]);
-    }*/
+    }
 }
 
-void acabaJogo(){
+void acabaJogo() {
     int i, fd;
     serv_clie j;
-    
+
 
     resultados.fim = 1;
     for (i = 0; i < TOTAL; i++) {
@@ -1492,51 +1490,67 @@ void acabaJogo(){
     }
 
     pthread_join(tarefa_bola, NULL);
-    
+
     usleep(100);
-    
-    for(i=0;i<clientes.tam;i++)
-    {
-        if(clientes.c[i].logado)
-        {
+
+    for (i = 0; i < clientes.tam; i++) {
+        if (clientes.c[i].logado) {
             char str[80];
             sprintf(str, "/tmp/ccc%d", clientes.c[i].id);
             fd = open(str, O_WRONLY);
-            
-            j.flag_campo=0;
-            j.flag_logado=0;
-            j.flag_stop=1;
+
+            j.flag_campo = 0;
+            j.flag_logado = 0;
+            j.flag_stop = 1;
+            j.resultados = resultados;
+
+            write(fd, &j, sizeof (serv_clie));
         }
     }
 }
 
-void * contar_seg(void * dados) {
-    int * tempo = (int *) dados;
-    serv_clie j;
-    j.flag_campo = 1;
-    j.xant = 99;
-    j.xnovo = 99;
-    j.yant = 99;
-    j.ynovo = 99;
+//void * contar_seg(void * dados) {
+//    int * tempo = (int *) dados;
+//    serv_clie j;
+//    j.flag_campo = 1;
+//    j.xant = 99;
+//    j.xnovo = 99;
+//    j.yant = 99;
+//    j.ynovo = 99;
+//
+//    while (tempo > 0) {
+//        resultados.tempo--;
+//        j.resultados = resultados;
+//
+//        atualizaCampo(&j);
+//        sleep(1);
+//        tempo--;
+//    }
+//    resultados.fim = 1;
+//}
 
-    while (tempo > 0) {
-        resultados.tempo--;
-        j.resultados = resultados;
+void trataSinal(int s) {
 
-        atualizaCampo(&j);
-        sleep(1);
-        tempo--;
+    if (s == SIGALRM) {
+        printf("\nACABOU TEMPO\n");
+        resultados.fim = 1;
+
+        int i;
+        for (i = 0; i < TOTAL; i++) {
+            pthread_join(JOG[0][i].thread, NULL);
+            pthread_join(JOG[1][i].thread, NULL);
+        }
+
+
+        //        acabaJogo();
+        return;
     }
-    resultados.fim = 1;
-}
 
-void acabar_jogo(int s) {
-    //if (s == SIGINT) {
+
+
     resultados.fim = 1;
     unlink(FIFO);
 
-    //pthread_join(tempo, NULL);
-    //}
     if (s == SIGINT) {
         int i;
         for (i = 0; i < clientes.tam; i++) {
@@ -1549,26 +1563,31 @@ void acabar_jogo(int s) {
     }
 }
 
-void shutdown(){
+void shutdown() {
     acabaJogo();
-    
+
     int i;
     char str[20];
-    
+
     resultados.fim = 1;
     pthread_join(tempo, NULL);
-    
-     for (i = 0; i < clientes.tam; i++) 
-     {
+
+    for (i = 0; i < clientes.tam; i++) {
         kill(clientes.c[i].id, SIGUSR1);
         sprintf(str, "/tmp/ccc%d", clientes.c[i].id);
         unlink(str);
-     }
-                    
-    free(JOG[0]);
-    free(JOG[1]);
+        free(&clientes.c[i]);
+    }
+
+
+    for (i = 0; i < TOTAL; i++) {
+        free(&JOG[0][i]);
+        free(&JOG[1][i]);
+    }
+    free(&JOG[0]);
+    free(&JOG[1]);
     free(JOG);
-    
+
     unlink(FIFO);
 }
 
@@ -1583,23 +1602,21 @@ int main(int argc, char** argv) {
 
 
     srand((unsigned int) time(NULL));
-    signal(SIGINT, SIG_IGN);
+    //    signal(SIGINT, SIG_IGN);
 
-    signal(SIGINT, acabar_jogo); //TODO: NAO ESTA A FUNCIONAR (ACHO)
-    signal(SIGALRM, acabar_jogo); //TODO: NAO ESTA A FUNCIONAR (ACHO)
+    signal(SIGINT, trataSinal); //TODO: NAO ESTA A FUNCIONAR (ACHO)
+    signal(SIGALRM, trataSinal); //TODO: NAO ESTA A FUNCIONAR (ACHO)
 
     char *primeiro;
     char *segundo;
     FILE *f;
 
-
-
-    int i, flag=0;
+    int i, flag = 0;
     char cmd[80];
-    
-    clientes.tam=0;
-    resultados.fim=1;
-    resultados.numClientes=clientes.tam;
+
+    clientes.tam = 0;
+    resultados.fim = 1;
+    resultados.numClientes = clientes.tam;
 
     //    const char* AVndefesa = getenv("NDEFESAS");
     //    const char* AVnavanc = getenv("NAVANCADOS");
@@ -1626,14 +1643,14 @@ int main(int argc, char** argv) {
         return 3;
     }*///TODO: Descomentar isto
 
-    
+
     pthread_t receber_cliente;
 
     pthread_create(&receber_cliente, NULL, &Func_receber_cliente, NULL);
-    
+
     mkfifo(FIFO, 0600);
-    
-    comecaJogo(); //----------------------------------------TODO: RETIRAR ISTO --------------------------------
+
+    //    comecaJogo(); //----------------------------------------TODO: RETIRAR ISTO --------------------------------
 
 
     do {
@@ -1664,7 +1681,7 @@ int main(int argc, char** argv) {
                 if (resultados.fim == 1) {
                     resultados.tempo = atoi(primeiro);
 
-                    //alarm(res.tempo);
+                    alarm(resultados.tempo);
 
                     comecaJogo();
                     // pthread_create(&tempo, NULL, &contar_seg, &res.tempo);
@@ -1683,9 +1700,8 @@ int main(int argc, char** argv) {
             case 2://USER
                 primeiro = strtok(NULL, " ");
                 segundo = strtok(NULL, " ");
-                
-                if(primeiro==NULL || segundo==NULL)
-                {
+
+                if (primeiro == NULL || segundo == NULL) {
                     printf("\nErro de sintaxe: user {username} {password}\n");
                     break;
                 }
@@ -1704,13 +1720,12 @@ int main(int argc, char** argv) {
 
                         printf("\nUtilizador ja existe\n");
                         fclose(f);
-                        flag=1;
+                        flag = 1;
                         break;
                     }
                 }
-                
-                if(!flag)
-                {
+
+                if (!flag) {
                     f = fopen(clientes.nome_ficheiro, "at");
 
                     if (!f) {
@@ -1736,26 +1751,22 @@ int main(int argc, char** argv) {
                 printf("\nEquipa a: %d - Equipa b: %d\n", resultados.res_eq1, resultados.res_eq2); //mostrar o resultado ao admin
                 break;
             case 5://RED
-                if(segundo==NULL)
-                {
+                if (segundo == NULL) {
                     printf("\nErro de sintaxe: red {username}\n");
                     break;
                 }
                 segundo = strtok(NULL, " ");
-                for (i = 0; i < clientes.tam; i++) 
-                {
-                    if (strcmp(clientes.c[i].username, segundo) == 0) 
-                    {
-                        clientes.c[i].logado=0;
-                        if(clientes.c[i].jogador!=NULL)
-                        {
-                            clientes.c[i].equi='-';
-                            clientes.c[i].jogador->humano=0;
+                for (i = 0; i < clientes.tam; i++) {
+                    if (strcmp(clientes.c[i].username, segundo) == 0) {
+                        clientes.c[i].logado = 0;
+                        if (clientes.c[i].jogador != NULL) {
+                            clientes.c[i].equi = '-';
+                            clientes.c[i].jogador->humano = 0;
                         }
                         strcpy(clientes.c[i].username, "");
                         kill(clientes.c[i].id, SIGUSR1);
                         resultados.numClientes--;
-                        
+
                         break;
                     }
                 }
